+++ src/server/newscripts/Pandaria/MogushanVault/boss_will_of_emperor.cpp
@@ -314,7 +314,7 @@ class boss_generic_imperator : public CreatureScript
                 DoZoneInCombat(me, 150.0f);
             }
 
-            void RegeneratePower(Powers power, int32 &value)
+            void RegeneratePower(Powers power, float &value)
             {
                 if (!me->isInCombat())
                 {
